Added TinyGPS Library

+++ Flight_Controller_V1.0/G_Control.ino
@@ -1,17 +1,19 @@
 void control(){
   
-    double gps_previous;
-    double imu_previous;
+    float gps_previous;
+    float imu_previous;
     
-   /* 
+   
    if(millis() - gps_previous > gps_period){
         position_control();
         gps_previous = millis();
     }
-    */
-
+    
+/*
     if(millis() - imu_previous > imu_period){
         attitude_control();
         imu_previous = millis();
     }
+    */
+   
 }
